Clean up yard documentation and move password class to its own file
