+++ README.md
@@ -4,7 +4,7 @@ This Ruby Gem provides FFI bindings, and a simplified interface, to the Argon2
 algorithm. [Argon2](https://github.com/P-H-C/phc-winner-argon2) is the official
 winner of the Password Hashing Competition, a several year project to identify a
 successor to bcrypt/PBKDF/scrypt methods of securely storing passwords. This is
-an independant project and not official from the PHC team.
+an independent project and not official from the PHC team.
 
 ![Build Status](https://github.com/technion/ruby-argon2/workflows/Test%20Suite/badge.svg)
 [![Code Climate](https://codeclimate.com/github/technion/ruby-argon2/badges/gpa.svg)](https://codeclimate.com/github/technion/ruby-argon2)
