+++ lib/argon2/version.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
-# Standard Gem version constant.
-
 module Argon2
+  ##
+  # Standard Gem version constant.
+  #
   VERSION = "2.0.3"
 end
